+++ crates/core/src/test_utils/fixtures.rs
@@ -1,11 +1,9 @@
 use futures::executor::block_on;
 use rstest::*;
 
-use crate::test_utils::deploy_helpers::{KakarotTestEnvironmentContext, TestContext};
+use crate::test_utils::deploy_helpers::KakarotTestEnvironmentContext;
 
 #[fixture]
-pub fn kakarot_test_env_ctx(
-    #[default(TestContext::Simple)] test_context: TestContext,
-) -> KakarotTestEnvironmentContext {
-    block_on(async { KakarotTestEnvironmentContext::new(test_context).await })
+pub fn kakarot_test_env_ctx() -> KakarotTestEnvironmentContext {
+    block_on(async { KakarotTestEnvironmentContext::from_dump_state().await })
 }