+++ crates/eth-rpc/tests/integration.rs
@@ -1,22 +1,24 @@
 #[cfg(test)]
 mod integration_tests {
     use std::str::FromStr;
-    use std::sync::Arc;
 
+    use dotenv::dotenv;
     use ethers::prelude::{Block as EthersBlock, Http as EthersHttp, H256 as EthersH256};
-    use kakarot_rpc::test_utils::setup_kakarot_rpc_integration_env;
-    use kakarot_rpc_core::test_utils::deploy_helpers::construct_kakarot_test_sequencer;
+    use kakarot_rpc::test_utils::start_kakarot_rpc_server;
+    use kakarot_rpc_core::test_utils::deploy_helpers::KakarotTestEnvironmentContext;
+    use kakarot_rpc_core::test_utils::fixtures::kakarot_test_env_ctx;
     use reth_primitives::U64;
+    use rstest::*;
 
+    #[rstest]
     #[tokio::test]
-    async fn test_get_block_rpc() {
-        // Setup the test sequencer
-        let starknet_test_sequencer = Arc::new(construct_kakarot_test_sequencer().await);
+    async fn test_get_block_rpc(kakarot_test_env_ctx: KakarotTestEnvironmentContext) {
+        // Load env
+        dotenv().ok();
 
-        // Deploy the Kakarot contracts and start the Kakarot RPC server
-        let (server_addr, server_handle) = setup_kakarot_rpc_integration_env(&starknet_test_sequencer)
-            .await
-            .expect("Error setting up Kakarot RPC server");
+        // Start the Kakarot RPC server
+        let (server_addr, server_handle) =
+            start_kakarot_rpc_server(&kakarot_test_env_ctx).await.expect("Error setting up Kakarot RPC server");
 
         // Run the test
         let provider = EthersHttp::from_str(format!("http://localhost:{}", server_addr.port()).as_ref()).unwrap();