+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to
 ### Added
 
 - ci: add `CHANGELOG.md` and enforce it is edited for each PR on `main`
+- test: update integration tests to use prepopulated Katana dumped state